requirements.txt
Django==3.2
djangorestframework==3.12.4
django-filter==2.4.0
